refactor: file splitting and refactoring
