+++ src/api/youtube/youtube.api.js
@@ -1,16 +1,13 @@
-import axios from 'axios';
+import api from './api';
 
-const VIDEO_BASE_URL = import.meta.env.VITE_YOUTUBE_SEARCH_VIDEO_URL;
 const YOUTUBE_KEY = import.meta.env.VITE_YOUTUBE_KEY;
 
-export const getVideoData = async (videoIds) => {
+export const getVideoData = async (category, paramObj) => {
+  paramObj.key = YOUTUBE_KEY;
+
   try {
-    const response = await axios.get(VIDEO_BASE_URL, {
-      params: {
-        part: 'snippet',
-        id: videoIds,
-        key: YOUTUBE_KEY
-      }
+    const response = await api.get(`/${category}`, {
+      params: paramObj
     });
 
     return response.data.items;
@@ -18,4 +15,29 @@ export const getVideoData = async (videoIds) => {
     alert('영상을 가져오는데 실패했습니다.');
     return;
   }
+};
+
+export const getChannelData = async (channelId) => {
+  const response = await api.get('/channels', {
+    params: {
+      part: 'snippet',
+      id: channelId
+    }
+  });
+
+  return response;
+};
+
+export const getSearchData = async (searchQuery) => {
+  const response = await api.get('/search', {
+    params: {
+      part: 'snippet',
+      maxResults: 50,
+      q: searchQuery || import.meta.env.VITE_YOUTUBE_PARAM,
+      order: 'relevance',
+      type: 'playlist'
+    }
+  });
+
+  return response;
 };
