+++ src/components/BackModal/BackModalStyle.jsx
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+export const BackDrop = styled.div`
+  background-color: rgba(0, 0, 0, 0.5);
+`;
