+++ src/components/HomePage/MainBanner.jsx
@@ -1,23 +1,10 @@
 import Slider from 'react-slick';
 import 'slick-carousel/slick/slick-theme.css';
 import 'slick-carousel/slick/slick.css';
-import styled from 'styled-components';
+import { BannerContainer, BannerImage } from './MainBannerStyle';
 import Banner1 from '/src/assets/banner1.png';
 import Banner2 from '/src/assets/banner2.png';
 import Banner3 from '/src/assets/banner3.png';
-
-const BannerContainer = styled.div`
-  margin-bottom: 20px;
-  z-index: 1;
-  width: 100vw;
-  margin-left: calc(-50vw + 50%);
-`;
-
-const BannerImage = styled.img`
-  width: 100%;
-  height: 400px;
-  object-fit: cover;
-`;
 
 const images = [Banner1, Banner2, Banner3];
