+++ src/components/HomePage/MainBannerStyle.jsx
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+export const BannerContainer = styled.div`
+  margin-bottom: 20px;
+  z-index: 1;
+  width: 100vw;
+  margin-left: calc(-50vw + 50%);
+`;
+
+export const BannerImage = styled.img`
+  width: 100%;
+  height: 400px;
+  object-fit: cover;
+`;
