--- src/pages/Mypage.jsx
@@ -1,9 +1,0 @@
-const MyPage = () => {
-  return (
-    <div>
-      <h1>My Page</h1>
-    </div>
-  );
-};
-
-export default MyPage;
