+++ src/pages/PlayListDetail/PlayListDetail.jsx
@@ -1,42 +1,16 @@
 import { useEffect, useState } from 'react';
-import { v4 as uuidv4 } from 'uuid';
-import playBtn from '../assets/playBtn.png';
-import supabase from '../supabase/config';
+import playBtn from '../../assets/playBtn.png';
 
 import { useQuery } from '@tanstack/react-query';
-import dayjs from 'dayjs';
-import timezone from 'dayjs/plugin/timezone';
-import utc from 'dayjs/plugin/utc';
 import { useNavigate, useParams } from 'react-router-dom';
-import styled from 'styled-components';
-import { fetchPlaylist } from '../api/playlistApi';
-import { EditComment } from '../components/EditComment/EditComment';
-import useLogStore from '../zustand/user-log';
-// 1. 코딩애플 유튜브 id
-// 2. id를 이용해서 관련 목록을 가지고 올 방법이 있는지 찾아보는 것
-// PLfLgtT94nNq0qTRunX9OEmUzQv4lI4pnP -> 코딩애플 유튜브 채널 id -> 재생목록 리스트 [{}, {}, ...] ->
-
-const Container = styled.div`
-  max-width: 700px;
-  margin: auto;
-`;
-
-const CourseList = styled.section`
-  width: 100%;
-  margin: 10px auto;
-  height: 40px;
-  border-top: 2px solid var(--lime-color);
-  border-bottom: 2px solid var(--lime-color);
-`;
-
-const Details = styled.div`
-  width: 100%;
-  height: auto;
-  font-size: 1rem;
-  line-height: 32px;
-  margin: 20px auto;
-  text-overflow: ellipsis;
-`;
+import { createComment, deleteComment, findAllComment, updateComment } from '../../api/supabase/comment.api';
+import { updateInsertUserData } from '../../api/supabase/user.api';
+import { getVideoData } from '../../api/youtube/youtube.api';
+import { EditComment } from '../../components/EditComment/EditComment';
+import { date } from '../../utils/common/date';
+import useRecentVideoStore from '../../zustand/recent-video';
+import useLogStore from '../../zustand/user-log';
+import { Container, CourseList, Details } from './PlayListDetailStyle';
 
 const Detail = () => {
   const navigate = useNavigate();
@@ -45,17 +19,12 @@ const Detail = () => {
   const [comments, setComments] = useState('');
   const [dropdownStates, setDropdownStates] = useState({});
   const { user } = useLogStore();
-  dayjs.extend(utc);
-  dayjs.extend(timezone);
+  const addRecentVideo = useRecentVideoStore((state) => state.addRecentVideo);
+  const initializeRecentVideos = useRecentVideoStore((state) => state.initializeRecentVideos);
 
   useEffect(() => {
     const fetchComments = async () => {
-      const { data: Comments, error } = await supabase.from('Comments').select('*').eq('playlistId', id);
-      if (error) {
-        console.error(error);
-        alert('supabase에서 데이터를 가져오는 중 오류가 발생했습니다.');
-        return;
-      }
+      const Comments = await findAllComment(id);
       setCommentsInfo(Comments);
     };
     fetchComments();
@@ -67,7 +36,11 @@ const Detail = () => {
     error
   } = useQuery({
     queryKey: ['playlist', id],
-    queryFn: () => fetchPlaylist(id)
+    queryFn: () =>
+      getVideoData('playlistItems', {
+        part: 'snippet',
+        playlistId: id
+      })
   });
 
   const handleSubmit = async (e) => {
@@ -75,33 +48,11 @@ const Detail = () => {
 
     if (!comments.trim()) return alert('댓글 내용을 입력해주세요');
     if (!user) return alert('로그인을 해주세요.');
-    const date = dayjs().tz(dayjs.tz.guess()).format('YYYY-MM-DD HH:mm:ss');
 
-    try {
-      const { data, error } = await supabase
-        .from('Comments')
-        .insert({
-          commentId: uuidv4(),
-          userId: user.id,
-          playlistId: id,
-          content: comments,
-          createdAt: date,
-          email: user.email
-        })
-        .select('*');
+    const data = await createComment(user, comments, id);
 
-      if (error) {
-        console.error('댓글 삽입 중 오류 발생:', error);
-        alert('댓글 작성 중 오류가 발생했습니다.');
-        return;
-      }
-
-      setCommentsInfo((prevComments) => [...prevComments, ...data]);
-      setComments('');
-    } catch (err) {
-      console.error('오류 발생:', err);
-      alert('오류가 발생했습니다.');
-    }
+    setCommentsInfo((prevComments) => [...prevComments, ...data]);
+    setComments('');
 
     alert('댓글 작성이 완료됐습니다!');
   };
@@ -114,14 +65,9 @@ const Detail = () => {
 
   const handleUpdate = async (commentIdToUpdate, comments) => {
     if (!user) return alert('로그인을 해주세요.');
-    const { error } = await supabase.from('Comments').update({ content: comments }).eq('commentId', commentIdToUpdate);
-    if (error) {
-      console.error('댓글 수정 중 오류 발생:', error);
-      alert('댓글 수정 중 오류가 발생했습니다.');
-      return;
-    } else {
-      alert('댓글이 수정되었습니다!');
-    }
+
+    await updateComment(comments, commentIdToUpdate);
+
     setCommentsInfo((currentComments) =>
       currentComments.map((comment) =>
         comment.commentId === commentIdToUpdate ? { ...comment, content: comments } : comment
@@ -134,12 +80,8 @@ const Detail = () => {
     if (!user) return alert('로그인을 해주세요.');
     const confirmed = confirm('정말로 댓글을 삭제하시겠습니까?');
     if (confirmed) {
-      const { error } = await supabase.from('Comments').delete().eq('commentId', commentIdToDelete).select();
-      if (error) {
-        console.error('댓글 삭제 중 오류 발생:', error);
-        alert('댓글 삭제 중 오류가 발생했습니다.');
-        return;
-      }
+      await deleteComment(commentIdToDelete);
+
       setCommentsInfo((currentComments) =>
         currentComments.filter((comment) => comment.commentId !== commentIdToDelete)
       );
@@ -151,6 +93,29 @@ const Detail = () => {
       ...prevStates,
       [commentId]: !prevStates[commentId]
     }));
+  };
+
+  const handleOnClickCard = async (videoId) => {
+    if (user) {
+      const { email } = user;
+      await initializeRecentVideos(email);
+      await addRecentVideo(videoId);
+
+      const updateRecent = useRecentVideoStore.getState().recent;
+
+      const recentString = JSON.stringify(updateRecent);
+      const updateUserData = {
+        email,
+        recentVideo: recentString,
+        social: user.app_metadata.provider,
+        createdAt: date,
+        updatedAt: date
+      };
+
+      updateInsertUserData(updateUserData);
+    }
+
+    navigate(`/watch/${videoId}`);
   };
 
   if (isLoading) return <div>loading...</div>;
@@ -175,7 +140,7 @@ const Detail = () => {
           <li
             className="d-flex-row border-box"
             key={item.id}
-            onClick={() => navigate(`/watch/${item.snippet.resourceId.videoId}`)}
+            onClick={() => handleOnClickCard(item.snippet.resourceId.videoId)}
           >
             <img src={playBtn} alt="플레이어 버튼" height="18px" />
             <span>{item.snippet.title}</span>
