+++ src/pages/UnknownPage/UnknownPageStyle.jsx
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  background-color: var(--white-color);
+  text-align: center;
+`;
