+++ src/pages/VideoPage/VideoPage.jsx
@@ -1,18 +1,5 @@
 import { useParams } from 'react-router-dom';
-import styled from 'styled-components';
-
-const VideoContainer = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-`;
-
-const Iframe = styled.iframe`
-  width: 80%;
-  height: 80%;
-  border: none;
-`;
+import { Iframe, VideoContainer } from './VideoPageStyle';
 
 const VideoPage = () => {
   const { videoId } = useParams();
