+++ src/pages/VideoPage/VideoPageStyle.jsx
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+export const VideoContainer = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+`;
+
+export const Iframe = styled.iframe`
+  width: 80%;
+  height: 80%;
+  border: none;
+`;
