+++ src/shared/Router.jsx
@@ -1,13 +1,13 @@
 import { createBrowserRouter } from 'react-router-dom';
 import Layout from '../Layout/Layout';
 import LogInRedirect from '../components/LogIn/LogInRedirect';
-import MyPage from '../pages/MyPage/MyPage';
-import PlayListDetail from '../pages/PlayListDetail';
-import TestPage from '../pages/TestPage/TestPage';
-import VideoPage from '../pages/VideoPage';
 import useCourses from '../hooks/useCourses';
-import HomePage from '../pages/HomePage';
-import UnknownPage from '../pages/UnknownPage';
+
+import HomePage from '../pages/HomePage/HomePage';
+import MyPage from '../pages/MyPage/MyPage';
+import PlayListDetail from '../pages/PlayListDetail/PlayListDetail';
+import UnknownPage from '../pages/UnknownPage/UnknownPage';
+import VideoPage from '../pages/VideoPage/VideoPage';
 import useLogStore from '../zustand/user-log';
 
 const ProtectRoute = ({ element }) => {
@@ -34,14 +34,6 @@ const router = createBrowserRouter([
       },
       {
         path: '/watch/:id',
-        element: <div>페이지 이동 확인</div>
-      },
-      {
-        path: '/test',
-        element: <TestPage />
-      },
-      {
-        path: '/video/:videoId',
         element: <VideoPage />
       },
       {
